backend/requirements.txt
# ChildFocus Backend — Requirements
# Matches your actual installed environment (Python 3.13, Windows 11)
# Install: pip install -r requirements.txt
#
# SYSTEM REQUIREMENT (install separately, not via pip):
#   ffmpeg >= 7.0  →  https://www.gyan.dev/ffmpeg/builds/
#   Verify: ffmpeg -version

# ── Web Framework ──────────────────────────────────────────────────────────────
Flask==3.1.3
Werkzeug==3.1.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
blinker==1.9.0
click==8.3.1

# ── Database ───────────────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.48
alembic==1.18.4
Mako==1.3.10
greenlet==3.3.2

# ── Video Downloading ──────────────────────────────────────────────────────────
yt-dlp>=2026.2.21

# ── Computer Vision ────────────────────────────────────────────────────────────
# NOTE: Keep numpy < 2.3 to satisfy both opencv packages
numpy>=2.0,<2.3
opencv-python-headless==4.13.0.92

# ── Audio Analysis ─────────────────────────────────────────────────────────────
librosa==0.11.0
soundfile==0.13.1
audioread==3.1.0
audioop-lts==0.2.2
soxr==1.0.0
llvmlite==0.46.0
numba==0.64.0
pooch==1.9.0
lazy_loader==0.4
msgpack==1.1.2
decorator==5.2.1

# ── Scientific Computing ───────────────────────────────────────────────────────
scipy==1.17.1
scikit-learn==1.8.0
joblib==1.5.3
threadpoolctl==3.6.0

# ── Image Processing ───────────────────────────────────────────────────────────
Pillow>=11.0.0

# ── HTTP / Networking ──────────────────────────────────────────────────────────
requests==2.32.5
urllib3==2.6.3
certifi>=2026.2.25
charset-normalizer==3.4.4
idna==3.11

# ── Environment & Utilities ────────────────────────────────────────────────────
python-dotenv>=1.2.2
cffi==2.0.0
pycparser==3.0
packaging==26.0
platformdirs==4.9.2
colorama==0.4.6
typing_extensions==4.15.0

# ── Standard Library Backports (Python 3.13) ───────────────────────────────────
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
